requirements.txt
fastapi
uvicorn
openai
llama-index 
llama-index-embeddings-huggingface
llama-index-utils-workflow
llama-index-embeddings-google-genai
llama-index-llms-google-genai
llama-index-vector-stores-pinecone
pyvis
llama-index-experimental
